pylint and pep8 correction

+++ executer.py
@@ -6,27 +6,39 @@ import sys
 
 
 class NonDirName(Exception):
-    pass
-
+    """
+    path is not path to directory
+    """
 
 class InvalidSyntax(Exception):
-    pass
+    """
+    Not correct syntax
+    """
 
 
 class InvalidArgument(Exception):
-    pass
+    """
+    Not correct argument
+    """
 
 
 class NonFileName(Exception):
-    pass
+
+    """
+    path is not path to file
+    """
 
 
 class InvalidDir(Exception):
-    pass
+    """
+    Not correct directory
+    """
 
 
 class CurrentDirDeleting(Exception):
-    pass
+    """
+    Deleting of working directory
+    """
 
 
 class Executer:
@@ -34,6 +46,11 @@ class Executer:
         pass
 
     def chdir(self, subcommand):
+        """
+        changes directory
+        """
+        if isinstance(self, int):
+            raise SyntaxError
 
         place = path.expanduser(subcommand[0])
         if os.path.isdir(place):
@@ -42,6 +59,11 @@ class Executer:
             raise NonDirName
 
     def listdir(self, aim_path: str, flags: str = ""):
+        """
+        list directory content
+        """
+        if isinstance(self, int):
+            raise SyntaxError
         if not path.isdir(aim_path):
             raise NonDirName
         aim_path += '/'
@@ -64,34 +86,54 @@ class Executer:
         result += "\n\033[37m"
         return result
 
-    def copy(self, input: str, output: str):
+    def copy(self, input_file: str, output: str):
+        """
+        copies file
+        """
+        if isinstance(self, int):
+            raise SyntaxError
         if not path.isdir(path.dirname(output)):
             if path.dirname(output) != "":
                 raise NonDirName
-        if not path.isfile(input):
+        if not path.isfile(input_file):
             raise InvalidSyntax
         if path.isdir(output):
             output += "/"
-            output += path.basename(input)
-        os.system("cp " + input + " " + output)
+            output += path.basename(input_file)
+        os.system("cp " + input_file + " " + output)
 
-    def move(self, input: str, output: str):
+    def move(self, input_file: str, output: str):
+        """
+        moves file
+        """
+        if isinstance(self, int):
+            raise SyntaxError
         if not path.isdir(path.dirname(output)):
             if path.dirname(output) != "":
                 raise NonDirName
-        if not path.isfile(input):
+        if not path.isfile(input_file):
             raise InvalidSyntax
         if path.isdir(output):
             output += "/"
-            output += path.basename(input)
-        os.system("mv " + input + ' ' + output)
+            output += path.basename(input_file)
+        os.system("mv " + input_file + ' ' + output)
 
     def remove_file(self, filename: str):
+        """
+        removes file
+        """
+        if isinstance(self, int):
+            raise SyntaxError
         if not path.isfile(filename):
             raise NonFileName
         os.remove(filename)
 
     def remove_directory(self, directory_name: str):
+        """
+        removes empty directory
+        """
+        if isinstance(self, int):
+            raise SyntaxError
         if not path.isdir(directory_name):
             raise NonDirName
         if os.listdir(directory_name) != []:
@@ -99,12 +141,22 @@ class Executer:
         os.removedirs(directory_name)
 
     def make_directory(self, dirname):
+        """
+        creates directory
+        """
+        if isinstance(self, int):
+            raise SyntaxError
         if not path.isdir(path.dirname(dirname)):
             raise NonDirName
 
         os.makedirs(dirname)
 
     def cat(self, filename):
+        """
+        prints file content
+        """
+        if isinstance(self, int):
+            raise SyntaxError
         if not path.isfile(filename):
             raise NonFileName
         file = open(filename, 'r')
@@ -113,6 +165,11 @@ class Executer:
         return ans
 
     def touch(self, filename):
+        """
+        creates empty file
+        """
+        if isinstance(self, int):
+            raise SyntaxError
         if not path.isdir(path.dirname(filename)):
             raise NonDirName
         if path.exists(filename):
@@ -121,6 +178,11 @@ class Executer:
         file.close()
 
     def subprocess(self, directory: str, subcommand: list, stdin, stdout, stderr):
+        """
+        runs subprocess
+        """
+        if isinstance(self, int):
+            raise SyntaxError
         os.chdir(directory)
         subprocess.run(subcommand, stdin=sys.stdin, stdout=sys.stdout, stderr=sys.stderr)
